requirements.txt
bleach
celery
concurrent-log-handler
cryptography
email-validator
flask
flask-bcrypt
flask-caching
flask-cors
flask-dance
flask-jwt-extended
flask-mailman
flask-migrate
flask-session
flask-socketio
flask-sqlalchemy
flask-talisman
gevent
gevent-websocket
marshmallow
marshmallow-sqlalchemy
opencv-python
psycopg2
pyotp
python-dateutil
pyyaml
qrcode
redis[hiredis]
requests
uwsgi